check if grades.csv file exists

+++ organizer.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-#!/bin/bash
 
 if [ ! -d "archive" ]
 then
@@ -9,6 +8,12 @@ fi
 timestamp=$(date +"%Y%m%d-%H%M%S")
 
 new_name="grades_$timestamp.csv"
+
+if [ ! -f "grades.csv" ]
+then
+    echo "Error: grades.csv does not exist."
+    exit 1
+fi
 
 mv grades.csv archive/$new_name
